pyproject.toml
[project]
name = "miniclaw"
version = "0.1.0"
description = "Educational personal server assistant with Docker sandboxing"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "groq>=0.4.0",
    "httpx>=0.27.0",
    "python-telegram-bot>=21.0",
    "python-dotenv>=1.0.0",
    "docker>=7.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.2.0",
]

[project.scripts]
miniclaw = "miniclaw.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
